Trebuchet: Squash Add Smartspace Widget for almost workspaces
Trebuchet: Add Smartspace Widget.

Since we removed QSB on top it looks empty at the fist glance after boot.
So, add "At a glance" widget provided by Google to the default workspace,
this allows the user to move it around or remove it as per their choice.

Trebuchet: Add smartspace widget to 2x2 and 3x3 grids
Change-Id: I4e3183db981be71c5430e8b389f32907534c1444